Perception: moved memory operation related util to common folder.
