+++ modules/perception/obstacle/camera/detector/common/BUILD
@@ -3,21 +3,6 @@ load("//tools:cpplint.bzl", "cpplint")
 package(default_visibility = ["//visibility:public"])
 
 cc_library(
-    name = "util",
-    srcs = ["util.cc"],
-    hdrs = ["util.h"],
-    deps = [
-        "//modules/common:log",
-        "@caffe//:lib",
-        "@cuda",
-        "@eigen",
-        "@opencv2//:core",
-        "@opencv2//:imgproc",
-        "@undistort",
-    ],
-)
-
-cc_library(
     name = "feature_extractor",
     srcs = ["feature_extractor.cc"],
     hdrs = ["feature_extractor.h"],
